pyproject.toml
[project]
name = "neural_cg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "gstools>=1.6.1",
    "hydra-core>=1.3.2",
    "ilupp>=1.0.2",
    "imageio>=2.37.0",
    "ipython>=9.0.2",
    "lightning>=2.5.0.post0",
    "loguru>=0.7.3",
    "meshio>=5.3.5",
    "mlflow>=2.20.3",
    "pandas>=2.2.3",
    "pymeshfix>=0.17.0",
    "pyqt6>=6.8.1",
    "seaborn>=0.13.2",
    "tetgen>=0.6.5",
    "torch==2.6.0",
    "torch-geometric>=2.6.1",
    "trimesh>=4.6.6",
]

# specify the code is in `neural_cg` folder:
[tool.setuptools]
packages = ["neural_cg"]
include-package-data = true

